requirements.txt
aiofiles==0.7.0
aiohttp==3.7.3
aiohttp-cors==0.7.0
aiosqlite==0.16.1
appdirs==1.4.4
APScheduler==3.7.0
asgiref==3.4.1
Async-OAuthlib==0.0.9
async-timeout==3.0.1
asyncpg==0.24.0
attrs==20.3.0
autopep8==1.5.7
Babel==2.9.1
better-profanity==0.7.0
black==21.7b0
blinker==1.4
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.5
chardet==3.0.4
charset-normalizer==2.0.4
click==7.1.2
colorama==0.4.4
colorlog==4.7.2
commonmark==0.9.1
cryptography==3.2
discord==1.7.3
discord-components==1.1.4
discord-ext-ipc==2.1.1
discord-py-slash-command==2.4.0
py-cord==1.7.3
dislash.py==1.4.6
Flask==2.0.1
Flask-Babel==1.0.0
Flask-Cors==3.0.10
Flask-Discord==0.1.64
fuzzywuzzy==0.18.0
gunicorn==20.1.0
h11==0.12.0
h2==4.0.0
hikari==2.0.0.dev100
hpack==4.0.0
httptools==0.3.0
Hypercorn==0.11.2
hyperframe==6.0.1
idna==2.10
iso8601==0.1.16
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
MouseInfo==0.1.3
multidict==5.1.0
mypy-extensions==0.4.3
mysql-connector-python==8.0.26
oauthlib==3.1.1
pathspec==0.9.0
pickle-mixin==1.0.2
Pillow==8.3.1
pprintpp==0.4.0
priority==2.0.0
protobuf==3.17.3
psutil==5.8.0
pure25519==0.0.1
PyAutoGUI==0.9.53
pycodestyle==2.7.0
pycord==0.0.1
pycparser==2.20
pygame==2.0.1
PyGetWindow==0.0.9
Pygments==2.10.0
PyJWT==1.7.1
PyMsgBox==1.0.9
PyMySQL==1.0.2
PyNaCl==1.4.0
pynpm==0.1.2
pyperclip==1.8.2
pypika-tortoise==0.1.1
PyRect==0.1.4
PyScreeze==0.1.27
python-dotenv==0.15.0
pytube==11.0.1
pytz==2021.1
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.3.0
rich==1.3.1
sanic==21.6.2
Sanic-Cors==1.0.0
sanic-routing==0.7.1
six==1.16.0
sr-api==1.1.3
tortoise-orm==0.17.6
waitress==1.4.3
wavelink==0.9.9
websocket-client==0.57.0
wsproto==1.0.0
yarl==1.6.3
youtube-dl==2021.6.6
